requirements.txt
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
google-api-core==2.24.0
google-auth==2.38.0
google-cloud-bigquery==3.29.0
google-cloud-bigquery-migration==0.11.12
google-cloud-core==2.4.1
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
mysql-connector-python==9.2.0
packaging==24.2
proto-plus==1.25.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9
six==1.17.0
urllib3==2.3.0
